requirements.txt
python-dotenv
discord.py[voice]
discord
spotipy
yt-dlp
ffmpeg